+++ apps/web/src/components/pages/math-exams-landing.tsx
@@ -18,7 +18,7 @@ import { Instance } from '@/fetcher/graphql-types/operations'
 import { breakpoints } from '@/helper/breakpoints'
 import { cn } from '@/helper/cn'
 import { serloDomain } from '@/helper/urls/serlo-domain'
-import type { RegionData } from '@/pages/mathe-pruefungen/[region]'
+import type { RegionData } from '@/pages/pruefungen/[region]'
 
 export function MathExamsLanding({ region: initRegion }: RegionData) {
   const [region, setRegion] = useState<SupportedRegion>(initRegion ?? 'bayern')
@@ -112,8 +112,6 @@ export function MathExamsLanding({ region: initRegion }: RegionData) {
 
         <section className="relative -left-side w-[calc(100%+32px)] bg-blueWave bg-100% pt-24 text-center">
           <ExamsCompleteList region={region} />
-
-          {/* Banner für Autor*innengewinnung? */}
         </section>
       </main>
       <FooterNew />
